requirements.txt
py-getch
nltk